Update BWC zip links (#625)
